apply comment.
